Add new module `linalg`
