+++ brainunit/linalg/__init__.py
@@ -13,22 +13,13 @@
 # limitations under the License.
 # ==============================================================================
 
-__version__ = "0.0.2"
+from ._linalg_change_unit import *
+from ._linalg_change_unit import __all__ as _linalg_change_unit_all
+from ._linalg_keep_unit import *
+from ._linalg_keep_unit import __all__ as _linalg_keep_unit_all
 
-from . import _matplotlib_compat
-from . import lax
-from . import math
-from . import linalg
-from ._base import *
-from ._base import __all__ as _base_all
-from ._celsius import *
-from ._celsius import __all__ as _celsius_all
-from ._unit_common import *
-from ._unit_common import __all__ as _common_all
-from ._unit_constants import *
-from ._unit_constants import __all__ as _constants_all
-from ._unit_shortcuts import *
-from ._unit_shortcuts import __all__ as _std_units_all
+__all__ = (_linalg_change_unit_all +
+              _linalg_keep_unit_all)
 
-__all__ = ['math'] + _common_all + _std_units_all + _constants_all + _base_all + _celsius_all
-del _common_all, _std_units_all, _constants_all, _base_all, _celsius_all, _matplotlib_compat
+del (_linalg_change_unit_all,
+     _linalg_keep_unit_all)
