+++ brainunit/math/_fun_keep_unit_test.py
@@ -36,10 +36,6 @@ fun_keep_unit_math_unary = [
     'nanmedian', 'nanmean', 'nanstd', 'diff', 'nan_to_num',
 ]
 
-fun_keep_unit_math_unary_linalg = [
-    'norm',
-]
-
 fun_accept_unitless_unary_can_return_quantity = [
     'round', 'around', 'round_', 'rint',
     'floor', 'ceil', 'trunc', 'fix',
@@ -622,27 +618,6 @@ class TestFunKeepUnit(parameterized.TestCase):
             result = bm_fun(q)
             expected = jnp_fun(jnp.array(value))
             assert_quantity(result, expected, unit=unit)
-
-    @parameterized.product(
-        value=[(1.0, 2.0), (1.23, 2.34, 3.45)],
-        unit=[second, meter]
-    )
-    def test_fun_keep_unit_math_unary_linalg(self, value, unit):
-        bm_fun_list = [getattr(bm, fun) for fun in fun_keep_unit_math_unary_linalg]
-        jnp_fun_list = [getattr(jnp.linalg, fun) for fun in fun_keep_unit_math_unary_linalg]
-
-        for bm_fun, jnp_fun in zip(bm_fun_list, jnp_fun_list):
-            print(f'fun: {bm_fun.__name__}')
-
-            result = bm_fun(jnp.array(value))
-            expected = jnp_fun(jnp.array(value))
-            assert_quantity(result, expected)
-
-            q = value * unit
-            result = bm_fun(q)
-            expected = jnp_fun(jnp.array(value))
-            assert_quantity(result, expected, unit=unit)
-
     @parameterized.product(
         value=[((1.0, 2.0), (3.0, 4.0)),
                ((1.23, 2.34, 3.45), (4.56, 5.67, 6.78))],
